+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This project <em>does not yet</em> adhere to [Semantic Versioning](https://semve
 - Upgraded to Node 24 LTS
 - OTU Smart Selector now sorts recent list by most recently used [#4865]
 - API endpoint for observation matrices only returns public ones instead of all
+- `Download formatted` modal now display italics in preview box [#1856]
 
 ### Fixed
 
@@ -40,6 +41,7 @@ This project <em>does not yet</em> adhere to [Semantic Versioning](https://semve
 - The Browse taxon name task bar is positioned incorrectly when the header bar is fixed.
 - Filter Source: The style selection in formatted download always applies the same style regardless of what is selected.
 
+[#1856]: https://github.com/SpeciesFileGroup/taxonworks/issues/1856
 [#4661]: https://github.com/SpeciesFileGroup/taxonworks/issues/4661
 [#4851]: https://github.com/SpeciesFileGroup/taxonpages/issues/4851
 [#4852]: https://github.com/SpeciesFileGroup/taxonpages/issues/4852